Added filter dropdown for UTM fields (#25422)
ref https://linear.app/ghost/issue/NY-736/
ref https://linear.app/ghost/issue/NY-738/
- removed UTM dropdown from Web tab (stats and posts apps)
- added stats filter component (behind UTM labs flag)
- wired up new filters, updated endpoints to support all values
- added filter sync logic between Web and Locations tabs for Audience
filters

It became clear that UTMs require a different approach when wanting to
analyze the data. Rather than seeing which values were used for a
specific page, it could often be more instructive to see where the user
went *when* they arrived with a certain set of values, as UTM fields are
all about identifying how users got to a particular site (e.g. ad or
email campaign).

We removed the old dropdown from the Web tab and created new filters
that allow you to search and filter via Post, Source, Audience, and the
various UTMs, as well as combine filters. For now, it's a more naive
search with support for a single value and 'is equal to' logic. We'll
look to expand this as we explore more use cases. We're less interested
in allowing a full data slicer and want to be more intentional about
providing what's sensible within the context.