Fixed S3Storage uploads to Google Cloud Storage (GCS)
The PutObjectCommand doesn't play with streams on GCS as there are some
incompatibilities with signatures.

By using `readFile` we can pull the whole file into a buffer in memory
and upload it in a single chunk - which fixes the errors. This does
increase the memory footprint with larger files, but it's simpler and it
unblocks us for now and we will monitor memory usage and switch to a
chunk based approach if it's necessary