Fixed S3Storage uploads to Google Cloud Storage
ref https://linear.app/ghost/issue/PRO-1538

The PutObjectCommand doesn't play well with streams on GCS as
there are some incompatibilities with signatures.

By using `readFile` we can pull the whole file into a buffer in memory
and upload it in a single chunk - which fixes the errors. This does
increase the memory footprint with larger files, but it unblocks e2e
testing for now, we're working on a chunked multi-part upload which
will supercede this.