Cleaned up analytics test data scripts (#25582)
no ref

We have had a lot of iterations on development setup for our
analytics/Tinybird configuration. We've recently improved the `yarn
dev:analytics` script with some other improvements to the `yarn dev`
script (in the form of dev:forward for now). This PR cleans up all the
various ways of generating tests data with the intended flow being:

- `yarn dev:analytics` to spin up dev server(s), including Tinybird
local
- `yarn data:analytics:generate` to build data based on the Ghost db
content; this could be seeded with `yarn reset:data` or our
data-generator script
- `yarn data:analytics:clear` to reset the data in Tinybird (truncates
the db)