Added `automated_emails` table (#25494)
ref
https://linear.app/ghost/issue/NY-770/create-a-database-migration-to-add-automated-emails-table

## What

This PR adds a new table to Ghost's database, called `automated_emails`.
It also adds a new Claude Code "skill," which provides Claude Code with
instructions on how to create a database migration using the slimer CLI
tool.

## Why

We are working on adding welcome emails for new members to Ghost.
Currently Ghost can be configured to send a welcome email to members as
soon as they sign up, but the content of the welcome email is hardcoded.

Our next step is to allow users to update the content in the welcome
email, including the subject and body of the email. This table is where
we'll store the content of the welcome email, and any other future
automated emails we may add to Ghost.