🌐 Added German translations for Ghost, Portal (#25338)
ref #23361

The German locale was lacking a couple of translations (mostly phrases
related to the new verification code for member logins). As a native
German speaker and freelance dev with German clients, it made sense for
me to add these translations to the Ghost repo, so that everyone could
profit from it.

---------

Co-authored-by: Cathy Sarisky <42299862+cathysarisky@users.noreply.github.com>