Reduced e2e worker count on CI to 1
ref https://linear.app/ghost/issue/BER-3096

We observed a lot of flakyness when running multiple workers on a single CI e2e test runner. This changeset makes test worker counts depend on an environment variable, and sets that variable to 1 on CI which reduces the worker count.

In order to maintain CI test run durations, we increased the shards in the e2e matrix to 4.