Updated Analytics Web structure (#25583)
ref
https://linear.app/ghost/issue/NY-810/general-improvements-in-analytics-to-work-better-with-the-new-filter

- This PR moves the Locations tab under the Web tab in site-wide
analytics. The main reason is that Locations is very much related to Web
traffic both in terms of data and functions (filtering). Now that we add
more filtering options for web related data it makes sense to merge
Locations under Web.
- This change also allows "click-through" filtering — users can click on
a Location, Source etc. to activate filtering by that value. With
Locations on a separate tab, this experience becomes very scattered /
impossible to do it in a nice way.
- Quite a few visual details were off in Analytics in general which are
all related to this change: spacings were too wide, some emphasis
(icons) was missing on the sub navigation, keyboard navigation was
missing for the filters component (for advanced users) etc.
- Again, related to the above, the current NavBar component was not
flexible enough to be used for the new filtering mechanism. The usage of
CSS grid makes it much easier to use it for the new filtering UX.

---------

Co-authored-by: Steve Larson <9larsons@gmail.com>