Removed E2E test for Analytics > Locations tab (#25616)
no issue

This tab has been removed from the analytics UI, so we no longer need
this test. It was able to be merged with a failing test because the E2E
tests are not currently taken into account by this repo's branch
protection, which I'll fix in a separate PR. For now, this gets the E2E
tests back into a passing state.