Refactored `HtmlEditor` component used in settings (#25638)
ref
https://linear.app/ghost/issue/NY-836/add-minimaleditor-component-to-admin-x-design-system

In the member emails section of settings, we need to allow the user to
edit the content of the welcome email. Initially we'll just include
basic text formatting, but soon we'll want to add additional cards, like
buttons for call-to-actions.

Currently this modal uses the HtmlEditor, which already existed and was
used in e.g. newsletter settings to edit the footer content. This worked
initially, but the HtmlEditor outputs HTML, not Lexical, which breaks
when we try to wire it up to the API (which expects lexical).

This PR:
- Creates a new `KoenigEditorBase` component with the common boilerplate
needed to launch an instance of the editor
- Updates `HtmlEditor` to be a thin wrapper around `KoenigEditorBase` to
add the HTML output functionality
- Exports the `KoenigEditorBase` from the design system so we can create
custom editor instances for other use-cases, i.e. for member welcome
emails