Added ability to toggle welcome emails on/off in settings (#25614)
ref
https://linear.app/ghost/issue/NY-829/add-ability-to-toggle-each-type-of-welcome-email-onoff

The Member Welcome Emails settings UI was previously static with no
backend integration. Toggle states were not persisted and emails could
not be created or managed. This connects the UI to the new
automated_emails API endpoint, enabling users to toggle welcome emails
on/off with state persisting across page refreshes. Different default
Lexical content is used for free vs paid member emails.

This doesn't yet allow the content to be edited; this will be included in a subsequent comment.