Added ability to edit welcome email content (#25618)
ref
https://linear.app/ghost/issue/NY-830/add-ability-to-edit-welcome-emails-subject-content-and-sender-info

The member welcome emails UI in settings currently allows the user to
toggle the emails on/off, but does not allow them to edit the content or
any other fields. This adds the ability to edit the content, subject,
and sender info of the welcome emails after enabling them.