Skipped flaky locations test (#25643)
no ref

This test is passing locally and we have redundant other tests checking
filters, so it's not critical that this test is sorted out at the
moment. We've had some flakiness due to slowness with the Tinybird event
ingestion that we can revisit at large vs. extending a timeout on this
test.