🐛 Fixed contributor profile editor navigation on close (#25629)
ref https://linear.app/ghost/issue/BER-3039

When a contributor attempted to close their profile page they'd be re-directed back to it. This was caused by competing rules and side effects in the Settings component and the Ember router. This was fixed by lifting the redirect into the Router, which removed the conflict and allowed contributors to navigate away from their profile page.