Fixed free welcome emails being sent to new paid members (#25664)
closes
https://linear.app/ghost/issue/NY-857/free-member-welcome-email-is-sent-to-new-paid-members

Currently if free welcome emails are enabled, and a new member signs up
for a paid subscription, they will receive the free welcome email in
error. This is because new paid members are initially created in the
`free` status, until the Stripe subscription is fully processed, at
which point they transition to `paid`.

This fixes the bug by opting to skip sending the welcome email if the
member has a `stripeCustomer` attached, which indicates that the member
at least _intends_ to become a paid member.

---------

Co-authored-by: Troy Ciesco <tmciesco@gmail.com>