Added date filters to mv_session_data Tinybird pipe
All our Tinybird endpoints use data filters, but these filters aren't
passed to the mv_session_data pipe, which backs all the endpoints. This
means each time the endpoints are called, we run the mv_session_data
pipe without any date filters, so it runs against all data in Tinybird.

This adds the data filters to the mv_session_data pipe to improve the
query performance of the endpoints.