🐛 Fixed Analytics > Growth MRR chart starting at zero at beginning of time range (#25694)
ref
https://linear.app/ghost/issue/NY-328/investigate-mrr-chart-showing-zero-for-blueshirt-banter

## Problem

In some cases, the MRR chart in Analytics > Growth will incorrectly
start at $0 for the very first data point.
<img width="1428" height="496" alt="Screenshot 2025-12-10 at 21 44
47@2x"
src="https://github.com/user-attachments/assets/f53a6abc-6f51-4abe-872f-694ca1efa7c7"
/>


## Cause

The `/stats/mrr` endpoint returns a sparse dataset, only including dates
in the response that had changes in MRR, so not every data point in the
chart is returned by the API. The frontend logic for calculating the
first data point was not accounting for this, and defaulting to $0.

## Fix

The fix is to check for this condition where there are missing data
points between the `dateFrom` parameter passed to the API (the first day
of the selected date range) and the first data point returned by the
API, and to fill in the earliest MRR value for this range.