Added materialized view and duplicated v2 Tinybird endpoints (#25719)
ref
https://linear.app/ghost/issue/NY-865/analytics-sources-not-populating-for-tangle-due-to-408-timeouts

## Problem

Tinybird endpoints are timing out in production for the sites with the
most data. The `mv_session_data` pipe, which almost all our endpoints
depend on, calculates aggregations at the session level at query time, which we
believe to be the biggest contributor to poor performance when querying
against large data sets.

## Fix

The solution to fix this is to convert `mv_session_data` from a pipe to
a materialized view, such that Tinybird will calculate these
aggregations at ingest time instead of at query time. As this is a
rather large change to the implementation, we've opted to create a
duplicate v2 pipeline rather than updating the existing pipeline in
place. This way we can validate the v2 pipeline in production against
real production data, without actually changing any of the user-facing
behavior, before cutting over to using the new implementation.

---

## Changes made

- Creates a new datasource for the materialized view for
`mv_session_data`
- Creates mv_session_data_v2 pipe to feed the materialized view
- Creates filtered_sessions_v2, which is backed by the new materialized
view
- Creates v2 of all endpoints, which are backed by the
filtered_sessions_v2 and the new materialized view
- Duplicates all test files for endpoints, without making any changes to
the test cases or expected results, to ensure that v1 and v2 of our
endpoints maintain the same behavior
- Adds ability to switch Ghost to use v2 endpoints via config, for the sake of
validating against live production data