Fixed Explore service MRR data (#25730)
ref https://linear.app/ghost/issue/ENG-2555/
- updated to only submit Connect data (real Stripe MRR)
- updated MRR query to use Stats service

The previous implementation had a couple bugs:

The MRR stats query was not actually part of the Members Stats service.
This was not caught in unit tests because it was all mocked. This query
is actually part of the more generic Stats service.

Additionally, we always submitted MRR data, including Stripe's 'test
mode' data. We only want to submit real MRR data.