🐛 Fixed attribution for email-only posts
ref https://linear.app/ghost/issue/NY-878/

When a free member clicks a link to view an email-only post, it takes them to a url like `/email/{uuid}`. We append the appropriate attribution to the url via query params to create the `ghost-history` entry for attributing to the email, however, this won't work unless the post is also published.

We have a gap where the attribution lookup via uuid is only finding published posts.

This PR allows us to find published and sent posts, so that we correctly attribute the url when directly navigating to an email-only post (which in general is 'unlisted').